1(a)/pdf-outline/pdf_outline_extractor/requirements.txt
pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.23.26
numpy==1.26.4
scikit-learn==1.5.0
PyMuPDF
pytesseract
pdf2image
scikit-learn
numpy
Pillow
